Updated input types
